requirements.txt
streamlit 
pandas 
plotly 
python-dotenv 
google-generativeai
wordcloud
